% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bancoTAI.R
\name{bancoTAI}
\alias{bancoTAI}
\title{Selecciona un banco de items al azar}
\usage{
bancoTAI(banco, diseno)
}
\arguments{
\item{banco}{un banco de items.}

\item{diseno}{Una matriz de dos columnas y tantas filas como modelos distintas se quieran extraer del banco}
}
\value{
Returna un sub banco de items muestreados del banco original.
}
\description{
Dado un banco de items, esta funcion extrae items generados con distintos modelos. La cantidad de items de cada tipo de diseno se
determina en la matriz diseno.
}
\examples{
bancoIT = genBancoDF(nomFijas = c('NombreIt','Modelo'),nparam = 15)
m2pl=matrix(c(0.2,1,-3,3),byrow = TRUE,ncol=2,nrow=2)
bancoIT=genitmodelo(100,"2PL",m2pl,bancoIT)
diseno=matrix(c(25,"2PL"),ncol=2,byrow = FALSE)
bancTeor = bancoTAI(bancoIT,diseno)
A ver si le encontramos la vuelat
}
