% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creathetas.R
\name{creathetas}
\alias{creathetas}
\title{Simula las habilidades}
\usage{
creathetas(distrth, ...)
}
\arguments{
\item{distrth}{Distribucion con la que se va a generar las habilidades}

\item{...}{Parametros especificos de la funcion que se elija para simulra las habilidades}
}
\value{
Un vector de habilidades de largo "n" donde el parametro n se define dentro de "..."
}
\description{
Dado un banco de items, esta funcion extrae items generados con distintos modelos. La cantidad de items de cada tipo de diseno se
determina en la matriz diseno.
}
\examples{
thetas=creathetas("Normal",n=1000,mean=0,sd=1)
}
