% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derISO_Info.R
\name{derISO_Info}
\alias{derISO_Info}
\title{Estima la derivada primera de la curva ISO y la funcion de informacion}
\usage{
derISO_Info(icciso, puntuni, nucleo, hd = NULL)
}
\arguments{
\item{icciso}{un vector con la estimacion isotonica}

\item{puntuni}{grilla donde esta evaluada la curva iso}

\item{nucleo}{alguna funcion de nucleo de dentro de las opciones en \strong{nucleosAll}}

\item{hd}{ventan para la estimacion, por defecto es NULL, en ese caso se estima segun formula dada, si se engresa un valor,
se toma ese.}
}
\value{
Una lista donde con:
\itemize{
\item deriv: -Derivada de la curva isotonica
\item Info: Funcion de informacion de la curva isotonica
}
}
\description{
Esta funcion estima para las curvas isotonicas la derivada primera
}
\examples{
\dontrun{
# No correr
# icciso = icciso1[['resfin']]
# puntuni = icciso1[['puntos']]
# resDer = derISO_Info(icciso,puntuni,nucleo = normal)
}
}
