% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERRYSES.R
\name{ERRYSES}
\alias{ERRYSES}
\title{Funcion para estimar el error y el sesgo de las estimaciones del TAI}
\usage{
ERRYSES(simData, grilla = seq(1:100)/100)
}
\arguments{
\item{simData}{matriz o data frame con los valores reales de habilidad y los estimados
que seria la salida de la funcion TAIgeneric.}

\item{grilla}{grilla de puntos de habilidad para estimar el RMSE y el sesgo local}
}
\value{
Returna una lista con:
\itemize{
\item \strong{RMGEN}: valores del sego y RMSE para la simulacion ingresada
\item \strong{RMBIATH}: estimaciones del RMSE y el sego para una grilla de valores de theta
}
}
\description{
Funcion para estimar el error y el sesgo de las estimaciones del TAI
}
\examples{
# No correr

}
