% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estFunG.R
\name{estFunG}
\alias{estFunG}
\title{AF para estimar las habilidades de los individuos}
\usage{
estFunG(
  resp,
  rotacion = c("varimax", "oblimin", "none"),
  corr = c("tet", "poly", "mixed", "cor"),
  ndims = 1
)
}
\arguments{
\item{resp}{Matriz de ceros y unos que representan las respuestas a los items}

\item{rotacion}{se indiqua que rotacion usar para el AF. Puede ser una de estas: varimax (defecto), oblimin o none (ninguna)}

\item{corr}{correlacion a utilizar: tet (tetracorica), poly (policorica), mixed, cor (Pearson)}

\item{ndims}{numero de dimensiones a estimar.}
}
\value{
Un data frame con las respuestas y al final nuevas columnas con las estimaciones de las habilidades resultantes del AF.
Una columna por cada dimension. Se identifican con la legra \emph{g} seguido del numero de la dimension.
}
\description{
Esta funcion utiliza analisis factorial (AF) para estimar las habilidades para cada fila de la matriz de respuesta
a un determinado conjunto de items.
Se pueden asumir una o mas dimensiones y utilizar distintos tipos de rotacion. Las habiliddes se estiman
segun las cargas factoriales de los individuos a los factores.
}
\examples{
\dontrun{
# Genero el banco vacio
# bancoIT = genBancoDF(nomFijas = c('NombreIt','Modelo'),nparam = 15)
# m2pl=matrix(c(0.2,1,-3,3),byrow = TRUE,ncol=2,nrow=2)
# bancoIT = genitmodelo(100,"2PL",m2pl,bancoIT)
# Especifico que voy a extraer del banco completo, bancoTAI, 25 items de dos parametros
# diseno=matrix(c(25,"2PL"),ncol=2,byrow = FALSE)
# bancTeor = bancoTAI(bancoIT,diseno)
# Genero las habilidades
# thetas=creathetas("Normal",n=1000,mean=0,sd=1)
# respu=GenResp(banco = bancTeor,thetas)
# Estimo las habilidades para una dimension
# respG = estFunG(resp = respu,rotacion = c('oblimin'),corr = c('tet'),ndims = 1)
}
}
