% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estFunG_Simple.R
\name{estFunG_Simple}
\alias{estFunG_Simple}
\title{Funcion para generar las funcions gs que se usan como
estadistico usando sumas simples o ponderadas sobre los items}
\usage{
estFunG_Simple(resp, grupos, pesos)
}
\arguments{
\item{resp}{Matriz de ceros y unos que representan las respuestas a los items}

\item{grupos}{es un list que tiene tantas componentes como g's (estadisticos de ordenacion) y en las filas los items que se usan en la suma ponderada o simple}

\item{pesos}{ponderaciones en la suma}
}
\value{
Un data frame con las respuestas y al final nuevas columnas con las estimaciones de las habilidades resultantes.
Una columna por cada dimension. Se identifican con la legra \emph{g} seguido del numero de la dimension.
}
\description{
Funcion para generar las funcions gs que se usan como
estadistico usando sumas simples o ponderadas sobre los items
}
\examples{
\dontrun{
# Ejemplo 1 dimension solo suma
# itdim1=colnames(respu)[1:80]
# itemsgs=list(dim1=itdim1)
# pesos=list(dim1=rep(1,80))
# respG=estFunG_Simple(respu,itemsgs,pesos)


Ejemplo para dos dimensiones
itdim1=colnames(respu)[1:50]
itdim2=colnames(respu)[45:80]
itemsgs=list(dim1=itdim1,dim2=itdim2)
pesos=list(dim1=runif(50),dim2=runif(36))
respG=estFunG_Simple(respu,itemsgs,pesos)
}
}
