% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estthetaNP.R
\name{estthetaNP}
\alias{estthetaNP}
\title{Etimacion NO parametrica de las habilidades}
\usage{
estthetaNP(scores, Dth = qnorm, ...)
}
\arguments{
\item{scores}{Data frame resultante con las respuestas a los items (no importan) y con columnas que contengan la estimacion de
los scores de los individuos por agun metodo, ya sea suma, cargas factoriales etc.}

\item{Dth}{distribucion que se le va a aplicar a los percentiles para transformarlos. Hay que poner alguna funcion del estilo
qnorm por ejemplo.}

\item{...}{Parametros adicianales que asociados a la funcion Dth. que se elija para hacer la transformacion.}
}
\value{
Un data frame con las mismas columnas que tenia \emph{scores}, pero se le agragan columnas identificadas como: seg (score g sin empates),
pcg (percentiles del score g sin empates) y dtg (es la transformacion aplicando la funcion Dth a pcg).
}
\description{
Esta funcion estima las habilidades de los individuos tomando como insumo algun estadistico (suma de scores, thetas estimados
con la funcion estFunG u otros).
ROmpe los scores con empates, calcula el percentil para cada individuo y realiza una transformacion llevando la uniforme al dominio
de alguna otra distribucion, por ejemplo la normal
}
\examples{
\dontrun{
# Genero el banco vacio
bancoIT = genBancoDF(nomFijas = c('NombreIt','Modelo'),nparam = 15)
m2pl=matrix(c(0.2,1,-3,3),byrow = TRUE,ncol=2,nrow=2)
bancoIT = genitmodelo(100,"2PL",m2pl,bancoIT)
# Especifico que voy a extraer del banco completo, bancoTAI, 25 items de dos parametros
diseno=matrix(c(25,"2PL"),ncol=2,byrow = FALSE)
bancTeor = bancoTAI(bancoIT,diseno)
# Genero las habilidades
thetas=creathetas("Normal",n=1000,mean=0,sd=1)
respu=GenResp(banco = bancTeor,thetas)
# Estimo las habilidades para una dimension usando AF
respG = estFunG(resp = respu,rotacion = c('oblimin'),corr = c('tet'),ndims = 1)
# Estimo los thetas de forma no parametrica y usando la transformacion a la distribucion normal
thetaest = estthetaNP(scores = respG,Dth = qnorm)
}
}
