% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funKL_int_NOPAR.R
\name{funKL_int_NOPAR}
\alias{funKL_int_NOPAR}
\title{Estima la funcion paroxima la integral en un entorno de un punto para estimar la funcion KL}
\usage{
funKL_int_NOPAR(iccNP, sepGrilla, entorno, jfijo)
}
\arguments{
\item{iccNP}{Estimacion de la ICC no parametrica en determinados puntos}

\item{sepGrilla}{Separecion entre los puntos de la grilla}

\item{entorno}{porcentaje que indica cuantos puntos tomar en el entorno del punto a estimar}

\item{jfijo}{posicion de la grilla donde se va a estimar la funcion}
}
\value{
Un vector con la funcion KL estimada en cada posicion \emph{jfijo}.
}
\description{
Esta funcion define el integrando para estimar la funcion de KL  para una ICC estimada de forma NO parámetrica
o la curva isotonica en un punto fijo de la grilla
}
\examples{
\dontrun{
# No correr
}
}
