% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funKL_int_PAR.R
\name{funKL_int_PAR}
\alias{funKL_int_PAR}
\title{FUncion para integrar en el calculo de la funcion de KL para las curvas paramétrica en un punto dado}
\usage{
funKL_int_PAR(arg, thj, a, b)
}
\arguments{
\item{arg}{es la variable que se va a integrar}

\item{thj}{es el punto the la grilla de theta donde se va a calcular KL}

\item{a, b}{son los parámetros de discriminacion y dificultad del modelo parametrico 2PL}
}
\value{
Un vector con la funcion KL estimada en cada thj.
}
\description{
Esta funcion estima es la que se integra en un entorno del punto para estimar KL
}
\examples{
# No correr
}
