% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genBancoDF.R
\name{genBancoDF}
\alias{genBancoDF}
\title{Genera un data frame vacio con un determinado numero de calumnas fijas de characters y un
conjunto de columnas numericas}
\usage{
genBancoDF(nomFijas = c("NombreIt", "Modelo"), nparam = 15)
}
\arguments{
\item{nomFijas}{es un vector con los nombres de las columnas que seran character. Los nombres de estas columnas
se llamaran igual lo que se ingrese en nomFijas}

\item{nparam}{entero que indica cuantas columnas numericas se van a cargar. Por defecto estas columnas se nombran con la letra
P y un numero secuencial.}
}
\value{
un data frame vacio.
}
\description{
Genera un data frame vacio con un determinado numero de calumnas fijas de characters y un
conjunto de columnas numericas
}
\examples{
bancoIT = genBancoDF(nomFijas = c('NombreIt','Modelo'),nparam = 15)
}
