% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genitmodelo.R
\name{genitmodelo}
\alias{genitmodelo}
\title{Genera el banco de items para diferentes modelos con sus respectivos parametros}
\usage{
genitmodelo(caitems, modelo, parametros, banco)
}
\arguments{
\item{caitems}{es la cantidad de items a ingresar en el banco.}

\item{modelo}{un modelo pertenciente a los siguientes por ahora 1PL = un parametro, 2PL = dos parametros, 3PL = tres parametros o CUB =
modelo cubico}

\item{parametros}{es una matriz que tiene dos columnas y tantas filas como parametros. Determina los limites inferiores y superiores para
simular los parametros.}

\item{banco}{es un data frame donde se van cargando los parametros y los nombres de los items. Este data frame es el generado por la funcion
genBancoDF}
}
\value{
El mismo data frame con los nombres de los items y los parametros agregados debajo de los que ya existian.
}
\description{
Genera el banco de items para diferentes modelos con sus respectivos parametros
}
\examples{
bancoIT = genBancoDF(nomFijas = c('NombreIt','Modelo'),nparam = 15)
m2pl=matrix(c(0.2,1,-3,3),byrow = TRUE,ncol=2,nrow=2)
bancoIT=genitmodelo(100,"2PL",m2pl,bancoIT)
}
