% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenResp.R
\name{GenResp}
\alias{GenResp}
\title{Genera respuestas dada la habilidad theta}
\usage{
GenResp(banco, thetas)
}
\arguments{
\item{banco}{es un data frame con tantas filas como items tenga el Banco, en este dataframe se indica el modelo y los parámetros del ítem.}

\item{thetas}{es un vector con habilidades simuladas con la funcion \emph{creathetas}.}
}
\value{
Retorna una matriz con tantas filas como habilidades (individuos) y tantas columnas como items haya en el banco.
}
\description{
Para un banco de items parametrico simula una matriz de ceros (respuesta incorrecta) y unos (respuesta correcta) para cada uno de los
valores del vector theta.
}
\examples{
\dontrun{
# Genero el banco vacio
bancoIT = genBancoDF(nomFijas = c('NombreIt','Modelo'),nparam = 15)
m2pl=matrix(c(0.2,1,-3,3),byrow = TRUE,ncol=2,nrow=2)
bancoIT = genitmodelo(100,"2PL",m2pl,bancoIT)
# Especifico que voy a extraer del banco completo, bancoTAI, 25 items de dos parametros
diseno=matrix(c(25,"2PL"),ncol=2,byrow = FALSE)
bancTeor = bancoTAI(bancoIT,diseno)
# Genero las habilidades
thetas=creathetas("Normal",n=1000,mean=0,sd=1)
respu=GenResp(banco = bancTeor,thetas)
}
}
