% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iccFun.R
\name{iccFun}
\alias{iccFun}
\title{Para una grilla de puntos calcula la curva caracteristica del item ICC}
\usage{
iccFun(a, b, c, grilla)
}
\arguments{
\item{a}{parametro de discriminacion del la ICC}

\item{b}{parametro de dificultad del la ICC}

\item{c}{parametro de pseudo azar de la ICC}

\item{grilla}{valores de theta donde se evalua la funcion}
}
\value{
Retorna un vector con la funcion evaluada en los puntos de la grilla
}
\description{
Para una grilla de puntos calcula la curva caracteristica del item ICC
}
\examples{
a = 1; b = 0; c = 0.1
grilla = dnorm(seq(0,1,0.01))
p = iccFun(a,b,c,grilla)
}
