% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icciso.R
\name{icciso}
\alias{icciso}
\title{Funcion para estimar las curvas isotonas en dimension 1}
\usage{
icciso(icc1, hd, thetaiso, nt, puntosicc, nucleod)
}
\arguments{
\item{icc1}{estimados de la ICC no parametrica icc1 y puntosicc deben tener la misma longuitudla y curva estimada en el intervalo \link{0,1}}

\item{hd}{ventana para estimacion de la densidad}

\item{thetaiso}{vector donde se calculara la curva isotona}

\item{nt}{puntos en la grilla}

\item{puntosicc}{indica los valores donde esta calculada la icc1}

\item{nucleod}{nucle a usar para estimar la densidad}
}
\value{
Returna un sub banco de items muestreados del banco original.
}
\description{
Funcion para estimar las curvas isotonas en dimension 1
}
\examples{
# No correr
#icciso1 = icciso(icc1,hd,theta,nt,puntosiso,nucleod)
}
