% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl_mat_NOpar.R
\name{kl_mat_NOpar}
\alias{kl_mat_NOpar}
\title{Estima la funcion de KL para un conjunto de items estimados de forma NO parametrica}
\usage{
kl_mat_NOpar(iccNP_mat, sepGrilla, entorno, puntuni, shiftFun = qnorm)
}
\arguments{
\item{iccNP_mat}{data frames con las estimaciones de las ICC no parametricas. Los nombres de las columnas tienen que ser los nombres
de los items.}

\item{sepGrilla}{distancia entre los puntos consecutivos de la grilla. Mismos que la funcion \strong{funKL_int_NOPAR}}

\item{entorno}{porcentaje que indica cuantos puntos tomar en el entorno del punto a estimar.
Mismos que la funcion \strong{funKL_int_NOPAR}}

\item{puntuni}{grilla en el intervalo 0 1 donde estan estimados los puntos de la funcion icc no parametrica}

\item{shiftFun}{es la funcion que lleva a grilla 0 1 a la grilla de la distribucion que se indique, por defecto
se pone la normal}
}
\value{
Un data frame con tantas columnas como items y filas como puntos de la grilla
con las estimaciones del KL
}
\description{
Dado una secuencia de items estimados de forma NO parametrica (puede ser la isotonica),
estima la funcion KL para cada uno.
}
\examples{
\dontrun{
# No correr
# KL_MAT_NP = kl_mat_NOpar(iccNP_mat,sepGrilla,entorno)
}
}
