% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl_mat_par.R
\name{kl_mat_par}
\alias{kl_mat_par}
\title{Estima la funcion de KL para un conjunto de items estimados de forma parametrica}
\usage{
kl_mat_par(paramsMIRT, grillaEval, distTrans = qnorm)
}
\arguments{
\item{paramsMIRT}{data frames con los parametros estimados para elmodelo parametrico. Los nombres
de las filas de este data rame tienen que ser los nomres de los items.}

\item{grillaEval}{grilla de puntos donde se va a estimar la funcion}

\item{distTrans}{distribucion para transformar la grilla en caso de que la grilla este definida en $\link{0,1}$.}
}
\value{
Un data frame con tantas columnas como items y filas como puntos de la grilla
con las estimaciones del KL
}
\description{
Dado un na secuencia de items estimados de forma parametrica, estima la funcion KL para cada uno
}
\examples{
# No correr
}
