% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KL_NOPAR.R
\name{KL_NOPAR}
\alias{KL_NOPAR}
\title{Estima la funcion de KL para las curvas NO paramétrica en una grilla de puntos}
\usage{
KL_NOPAR(iccNP, sepGrilla, entorno)
}
\arguments{
\item{iccNP}{vecor con las estimaciones en una grilla}

\item{sepGrilla}{separacion entre puntos de la grilla}

\item{entorno}{porcentaje de puntos para tomar en el entorno}
}
\value{
Un vector con la funcion KL estimada en cada punto de la grilla.
}
\description{
Esta funcion estima la funcion de KL para para una ICC estimada de forma NO parámetrica en una grilla d epuntos
}
\examples{
\dontrun{
# No correr
}
}
