% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KL_PAR.R
\name{KL_PAR}
\alias{KL_PAR}
\title{Estima la funcion de KL para las curvas paramétrica en una grilla d epuntos}
\usage{
KL_PAR(nomItem, paramsMIRT, grillaEval, distTrans = qnorm)
}
\arguments{
\item{nomItem}{string con el nombre del item a estimar}

\item{paramsMIRT}{data frames con los parametros estimados para elmodelo parametrico}

\item{grillaEval}{grilla de puntos donde se va a estimar la funcion}

\item{distTrans}{distribucion para transformar la grilla en caso de que la grilla este definida en $\link{0,1}$.}
}
\value{
Un vector con la funcion KL estimada en cada punto de la grilla.
}
\description{
Esta funcion estima la funcion de KL para para una ICC estimada de forma parámetrica en una grilla d epuntos
}
\examples{
\dontrun{
# No correr
}
}
