% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucleosAll.R
\name{normal}
\alias{normal}
\title{Condunto de nucleos unidimensionales y multidimensionelaes}
\usage{
normal(x, h, th)
}
\arguments{
\item{x}{vector al que se le aplica el nucleo}

\item{h}{amplitud de la ventana.}

\item{th}{puntos donde se tiene valores de la funcion.}

\item{sigma}{matriz de varianza y covarianza para el caso de el nucleo \emph{multinorm}}
}
\value{
un vector de largo \emph{x} con el resultado de la aplicacion del nucleo
}
\description{
Los nucleos multidimensionales son: Normal multiplicativo (multinorm), Nucleo Epanechnikov esferico (epaesf),
Normal multivariada (multinorm)
}
\examples{
set.seed(1234)
xeval = seq(0,1,0.01)
h = 0.05
th = runif(1000)
res = normal(x = xeval,h = h,th = th)
}
