% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nucleosAll.R
\name{nucleosAll}
\alias{nucleosAll}
\alias{normal}
\alias{epa}
\alias{acuepa}
\alias{forme}
\alias{acunif}
\alias{trian}
\alias{acutrian}
\alias{Biwei}
\alias{acubiw}
\alias{Triwei}
\alias{acutri}
\alias{Trcubo}
\alias{acucub}
\alias{Cos}
\alias{acucos}
\alias{Logis}
\alias{aculogis}
\alias{Silver}
\alias{Sigmo}
\alias{acusig}
\alias{multinorm}
\alias{epaesf}
\alias{normmult}
\title{Condunto de nucleos unidimensionales y multidimensionelaes}
\usage{
normal(x, h, th)

epa(x)

acuepa(x)

forme(x)

acunif(x)

trian(x)

acutrian(x)

Biwei(x)

acubiw(x)

Triwei(x)

acutri(x)

Trcubo(x)

acucub(x)

Cos(x)

acucos(x)

Logis(x)

aculogis(x)

Silver(x)

Sigmo(x)

acusig(x)

multinorm(x, sigma)

epaesf(x)

normmult(x)
}
\arguments{
\item{x}{vector al que se le aplica el nucleo}

\item{h}{amplitud de la ventana.}

\item{th}{puntos donde se tiene valores de la funcion.}

\item{sigma}{matriz de varianza y covarianza para el caso de el nucleo \emph{multinorm}}
}
\value{
un vector de largo \emph{x} con el resultado de la aplicacion del nucleo
}
\description{
Los nucleos multidimensionales son: Normal multiplicativo (multinorm), Nucleo Epanechnikov esferico (epaesf),
Normal multivariada (multinorm)
}
\examples{
set.seed(1234)
xeval = seq(0,1,0.01)
h = 0.05
th = runif(1000)
res = normal(x = xeval,h = h,th = th)
}
