% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TAIgeneric.R
\name{TAIgeneric}
\alias{TAIgeneric}
\title{Funcion para simular la aplicacion de un TAI para items NO parametricos}
\usage{
TAIgeneric(
  sujtai,
  epsilon,
  minit,
  maxit,
  curvaNOPAR,
  parametros,
  itemsSelec = c("InfoIso", "KL", "ESH", "Random"),
  matrizSelect = MINOPARAUX,
  seqTheta
)
}
\arguments{
\item{sujtai}{Habilidades verdaderas de los estudiantes con distribucion normal}

\item{epsilon}{error maximo en la estimacion con respecto a la anterior}

\item{minit}{minima cantidad de items a aplicar en el TAI}

\item{maxit}{maxima contidad de items a aplicar}

\item{curvaNOPAR}{curvas estimadas de forma no parametrica (isotonicas o por regresion)}

\item{parametros}{parametros verdaderos de los items}

\item{itemsSelec}{Metodo utilizado para seleccionar los items:
\itemize{
\item \emph{InfoIso}: pseudo informacion
\item \emph{KL}: Kullback Leiber
\item \emph{ESH}: Esperanza de la entropia de shannon
\item \emph{random}: Seleccion aleatoria
}}

\item{matrizSelect}{en los casos de usar el metodo de seleccion InfoIso o KL
aqui se pone la matriz de informacion o de KL para todos los items en cada valor de la grilla}

\item{seqTheta}{Es la grilla de theta donde estan calculadas las matrices de info y KL}
}
\value{
Returna una matriz con el verdadero valor de theta, el estimado,
el error cometido y el total de items aplicado.
}
\description{
Funcion para simular la aplicacion de un TAI para items NO parametricos
}
\examples{
\dontrun{
# No correr
}

}
