% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ventana1D.R
\name{ventana1D}
\alias{ventana1D}
\title{Estima las ventanas para items unidimensionales}
\usage{
ventana1D(
  items,
  th_use = "pcg",
<<<<<<< HEAD
  test,
=======
  test = NULL,
>>>>>>> 9c6b2677dbbc5c868d036bfefe12c12dfbcabd62
  nucleodes = "gaussian",
  muestra = "TODO"
)
}
\arguments{
\item{items}{Puede ser un vector de strings con los nombres de los items o las posiciones.}

\item{th_use}{Referencia a la estimacion de los thetas a usar del data frame proveniente de la funcion
\emph{estthetaNP}, puede tomar los valores: pcg (usa los percentiles) o dtg (usa la tranformacion a alguna distribucion)}

\item{test}{Data frame o matriz con las respuestas de los individuos y las estimaciones de las habilidades. Salida
de la funcion \emph{estthetaNP}.}

\item{nucleodes}{Es es la descripción del Kernel utilizado.
Los valores permitidos son  gaussian, epanechnikov, o uniform}

\item{muestra}{cantidad de casos tomados para la estimacion, por defecto son todos.}
}
\value{
Una vector con los anchos de ventana vantanas
}
\description{
Esta funcion utiliza la libreria np
}
\examples{
<<<<<<< HEAD
\dontrun{
=======

>>>>>>> 9c6b2677dbbc5c868d036bfefe12c12dfbcabd62
# Genero el banco vacio
bancoIT = genBancoDF(nomFijas = c('NombreIt','Modelo'),nparam = 15)
m2pl=matrix(c(0.2,1,-3,3),byrow = TRUE,ncol=2,nrow=2)
bancoIT = genitmodelo(100,"2PL",m2pl,bancoIT)
# Especifico que voy a extraer del banco completo, bancoTAI, 25 items de dos parametros
diseno=matrix(c(25,"2PL"),ncol=2,byrow = FALSE)
bancTeor = bancoTAI(bancoIT,diseno)
# Genero las habilidades
thetas=creathetas("Normal",n=1000,mean=0,sd=1)
respu=GenResp(banco = bancTeor,thetas)
# Estimo las habilidades para una dimension usando AF
respG = estFunG(resp = respu,rotacion = c('oblimin'),corr = c('tet'),ndims = 1)
# Estimo los thetas de forma no parametrica y usando la transformacion a la distribucion normal
thetaest = estthetaNP(scores = respG,Dth = qnorm)

# Estima la ventana para un item especifico
<<<<<<< HEAD
h=ventana1D(items = 1,th_use = 'dtg',test = thetaest,nucleodes="gaussian",1000)
}
=======
# h=ventana1D(items = 1,th_use = 'dtg',test = thetaest,nucleodes="gaussian",1000)
>>>>>>> 9c6b2677dbbc5c868d036bfefe12c12dfbcabd62
}
